veridical design - bootstrap HTML output format

Format for converting from R Markdown to an HTML document using the veridical design theme. The veridical theme is a wrapper around the vmodern design theme from the `vthemes` R package with the difference that the veridical theme is tailored specifically for the PCS lab notebook and thus adds interactive capabilities such as clickable checkboxes, editable textboxes, and collapsible tips. Note that the vmodern design theme and hence the veridical design theme are largely based upon the material design theme from the `rmdformats` R package. See https://github.com/juba/rmdformats for the source code.

veridical(
  fig_width = 10,
  fig_height = 8,
  number_sections = FALSE,
  code_folding = "hide",
  code_download = TRUE,
  use_bookdown = TRUE,
  includes = NULL,
  css = NULL,
  fig_caption = TRUE,
  highlight = "kate",
  lightbox = TRUE,
  thumbnails = TRUE,
  gallery = FALSE,
  cards = TRUE,
  pandoc_args = NULL,
  md_extensions = NULL,
  mathjax = "rmdformats",
  ...
)

Arguments

fig_width

Default width (in inches) for figures

fig_height

Default width (in inches) for figures

number_sections

TRUE to number section headings

code_folding

Enable document readers to toggle the display of R code chunks. Specify "none" to display all code chunks (assuming they were knit with echo = TRUE). Specify "hide" to hide all R code chunks by default (users can show hidden code chunks either individually or document-wide). Specify "show" to show all R code chunks by default.

code_download

Embed the Rmd source code within the document and provide a link that can be used by readers to download the code.

use_bookdown

if TRUE, uses html_document2 instead of html_document, thus providing numbered sections and cross references

includes

Named list of additional content to include within the document (typically created using the includes function).

css

CSS and/or Sass files to include. Files with an extension of .sass or .scss are compiled to CSS via sass::sass(). Also, if theme is a bslib::bs_theme() object, Sass code may reference the relevant Bootstrap Sass variables, functions, mixins, etc.

fig_caption

TRUE to render figures with captions

highlight

Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "textmate". Pass NULL to prevent syntax highlighting.

lightbox

if TRUE, add lightbox effect to content images

thumbnails

if TRUE display content images as thumbnails

gallery

if TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display

cards

if TRUE, sections will be presented as distinct and animated cards

pandoc_args

arguments passed to the pandoc_args argument of rmarkdown html_document

md_extensions

arguments passed to the md_extensions argument of rmarkdown html_document

mathjax

set to NULL to disable Mathjax insertion

...

Additional function arguments passed to `vthemes::vmodern()`.

Value

R Markdown output format to pass to render

Details

JavaScript and CSS taken and adapted from the Material design theme for Bootstrap 3 project : https://github.com/FezVrasta/bootstrap-material-design.